+++ client/src/main/java/replicant/ReplicantContext.java
@@ -34,11 +34,6 @@ public final class ReplicantContext
   @Nullable
   private final SpyImpl _spy;
   /**
-   * Support infrastructure for application events.
-   */
-  @Nullable
-  private final ApplicationEventBroker _eventBroker;
-  /**
    * Support infrastructure for change broker.
    */
   @Nullable
@@ -60,7 +55,6 @@ public final class ReplicantContext
     _validator = Validator.create( Replicant.areZonesEnabled() ? this : null );
     _schemaService = SchemaService.create();
     _spy = Replicant.areSpiesEnabled() ? new SpyImpl() : null;
-    _eventBroker = Replicant.areEventsEnabled() ? new ApplicationEventBroker() : null;
     _changeBroker = Replicant.isChangeBrokerEnabled() ? new EntityChangeBroker() : null;
   }
 
@@ -275,35 +269,6 @@ public final class ReplicantContext
     }
     assert null != _spy;
     return _spy;
-  }
-
-  /**
-   * Return true if application events will be propagated.
-   * This means events are enabled and there is at least one application event handler present.
-   *
-   * @return true if application events will be propagated, false otherwise.
-   */
-  boolean willPropagateApplicationEvents()
-  {
-    return Replicant.areEventsEnabled() && getEventBroker().willPropagateApplicationEvents();
-  }
-
-  /**
-   * Return the event broker associated with context.
-   * This method should not be invoked unless {@link Replicant#areEventsEnabled()} returns true.
-   *
-   * @return the event broker associated with context.
-   */
-  @Nonnull
-  public ApplicationEventBroker getEventBroker()
-  {
-    if ( Replicant.shouldCheckApiInvariants() )
-    {
-      apiInvariant( Replicant::areEventsEnabled,
-                    () -> "Replicant-0092: Attempting to get ApplicationEventBroker but events are not enabled." );
-    }
-    assert null != _eventBroker;
-    return _eventBroker;
   }
 
   /**