+++ client/src/main/java/replicant/events/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Notifications that appear to support event driven applications.
- * Classes in this package are only used to support legacy applications and should not be used for new applications.
- */
-package replicant.events;